backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Configuration
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring & Observability
prometheus-client==0.19.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
tiktoken==0.5.1

# Vector Search & AI
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.0

# Async Processing
celery==5.3.4
redis==5.0.1

